+++ tomcat/src/test/java/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -16,7 +16,7 @@ class Http11ProcessorTest {
     void process() {
         // given
         final var socket = new StubSocket();
-        final var processor = new Http11Processor(socket);
+        final var processor = new Http11Processor(socket, new StaticFileLoader());
 
         // when
         processor.process(socket);
@@ -43,7 +43,7 @@ class Http11ProcessorTest {
                 "");
 
         final var socket = new StubSocket(httpRequest);
-        final Http11Processor processor = new Http11Processor(socket);
+        final Http11Processor processor = new Http11Processor(socket, new StaticFileLoader());
 
         // when
         processor.process(socket);
